Seperate osm-p2p creation and API pieces.
This will make it easier to write unit tests that don't rely on React Native
being present.
