+++ app/screens/Observations/map.js
@@ -10,7 +10,7 @@ import {
 import Mapbox, { MapView } from "react-native-mapbox-gl";
 import Icon from "react-native-vector-icons/MaterialIcons";
 
-import osmp2p from "../../lib/osm-p2p";
+import osmp2p from "../../lib/create-osm-p2p";
 import { Header, SideMenu, Text } from "../../components";
 import { baseStyles, colors } from "../../styles";
